+++ .idea/libraries/lib.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="lib">
+    <CLASSES>
+      <root url="file://$PROJECT_DIR$/../../../../JAVAFX/javafx-sdk-14.0.2.1/lib" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="file://$PROJECT_DIR$/../../../../JAVAFX/javafx-sdk-14.0.2.1/lib" />
+    </SOURCES>
+    <jarDirectory url="file://$PROJECT_DIR$/../../../../JAVAFX/javafx-sdk-14.0.2.1/lib" recursive="false" />
+    <jarDirectory url="file://$PROJECT_DIR$/../../../../JAVAFX/javafx-sdk-14.0.2.1/lib" recursive="false" type="SOURCES" />
+  </library>
+</component>
