+++ src/com/company/app/Main.java
@@ -1,0 +1,26 @@
+package com.company.app;
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Main extends Application {
+
+    public static void main(String[] args) {
+/*
+        Kostka kostka = new Kostka(6);
+
+        new HraciPlocha(40, 4, 4, kostka);*/
+
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage stage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
+        stage.setTitle("Hello World");
+        stage.setScene(new Scene(root, 300, 275));
+        stage.show();
+    }
+}
