+++ src/com/company/model/Barva.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 
 public class Barva {
     private  int startpole;
