+++ src/com/company/model/CilDomecekTest.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 
 import org.junit.jupiter.api.Test;
