+++ src/com/company/model/Figurka.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 
 public class Figurka {
