+++ src/com/company/model/FigurkaTest.java
@@ -1,12 +1,12 @@
-package com.company;
+package com.company.model;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class KostkaTest {
+class FigurkaTest {
 
     @Test
-    void hod() {
-    }
+    void getBarva() {
+    };;
 }
