+++ src/com/company/model/HraciPlochaTest.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 
 import org.junit.jupiter.api.Test;
