+++ src/com/company/model/Startdomecek.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 import java.util.*;
 public class Startdomecek {
