+++ src/com/company/model/StartdomecekTest.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.model;
 
 import org.junit.jupiter.api.Test;
