+++ src/com/company/wiew/MainController.java
@@ -1,0 +1,4 @@
+package com.company.wiew;
+
+public class MainController {
+}
