+++ src/com/company/wiew/SceneBuilder.fxml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.Cursor?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuBar?>
+<?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.Pane?>
+
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #aaa;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
+   <children>
+      <Button layoutX="196.0" layoutY="373.0" mnemonicParsing="false" text="nasadit" />
+      <Button layoutX="145.0" layoutY="373.0" mnemonicParsing="false" text="kostka" />
+      <Button layoutX="250.0" layoutY="373.0" mnemonicParsing="false" text="posunuti" />
+      <Pane layoutX="14.0" layoutY="14.0" prefHeight="350.0" prefWidth="430.0" style="-fx-background-color: #bbb;">
+         <cursor>
+            <Cursor fx:constant="DEFAULT" />
+         </cursor>
+         <opaqueInsets>
+            <Insets />
+         </opaqueInsets>
+         <children>
+            <Button layoutX="190.0" layoutY="150.0" mnemonicParsing="false" prefHeight="50.0" prefWidth="50.0" text="num" />
+         </children>
+      </Pane>
+      <MenuBar layoutX="445.0" prefHeight="26.0" prefWidth="155.0" AnchorPane.rightAnchor="0.0">
+        <menus>
+          <Menu mnemonicParsing="false" text="File">
+            <items>
+              <MenuItem mnemonicParsing="false" text="Close" />
+            </items>
+          </Menu>
+          <Menu mnemonicParsing="false" text="Edit">
+            <items>
+              <MenuItem mnemonicParsing="false" text="Delete" />
+            </items>
+          </Menu>
+          <Menu mnemonicParsing="false" text="Help">
+            <items>
+              <MenuItem mnemonicParsing="false" text="About" />
+            </items>
+          </Menu>
+        </menus>
+      </MenuBar>
+      <ListView layoutX="445.0" layoutY="26.0" prefHeight="374.0" prefWidth="155.0" style="-fx-background-color: #ddd;" />
+      <TableView layoutX="445.0" layoutY="35.0" prefHeight="90.0" prefWidth="155.0">
+        <columns>
+          <TableColumn prefWidth="155.0" sortType="DESCENDING" style="-fx-background-color: #f00;" text="Hrac 1" />
+        </columns>
+      </TableView>
+      <TableView layoutX="445.0" layoutY="126.0" prefHeight="90.0" prefWidth="155.0">
+         <columns>
+            <TableColumn prefWidth="155.0" sortType="DESCENDING" style="-fx-background-color: #00f;" text="Hrac 2" />
+         </columns>
+      </TableView>
+      <TableView layoutX="445.0" layoutY="217.0" prefHeight="90.0" prefWidth="155.0">
+         <columns>
+            <TableColumn prefWidth="155.0" sortType="DESCENDING" style="-fx-background-color: #ff0;" text="Hrac 3" />
+         </columns>
+      </TableView>
+      <TableView layoutX="445.0" layoutY="308.0" prefHeight="90.0" prefWidth="155.0">
+         <columns>
+            <TableColumn prefWidth="155.0" sortType="DESCENDING" style="-fx-background-color: #0f0;" text="Hrac 4" />
+         </columns>
+      </TableView>
+   </children>
+</AnchorPane>
